backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
asyncpg>=0.29.0
aiofiles>=23.2.0
Pillow>=10.2.0
boto3>=1.34.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
